refactor: reorganize package structure by moving entities to a unified package
