+++ src/main/java/org/thingai/app/scoringservice/handler/EventHandler.java
@@ -3,8 +3,8 @@ package org.thingai.app.scoringservice.handler;
 import org.thingai.app.scoringservice.callback.EventHandlerCallback;
 import org.thingai.app.scoringservice.callback.RequestCallback;
 import org.thingai.app.scoringservice.define.ErrorCode;
-import org.thingai.app.scoringservice.entity.config.DbMapEntity;
-import org.thingai.app.scoringservice.entity.event.Event;
+import org.thingai.app.scoringservice.entity.DbMapEntity;
+import org.thingai.app.scoringservice.entity.Event;
 import org.thingai.app.scoringservice.repository.LocalRepository;
 import org.thingai.base.dao.exceptions.DaoException;
 import org.thingai.base.log.ILog;
