+++ src/main/java/org/thingai/app/scoringservice/strategy/IRankingStrategy.java
@@ -1,7 +1,9 @@
-package org.thingai.app.scoringservice.entity.ranking;
+package org.thingai.app.scoringservice.strategy;
 
 import org.thingai.app.scoringservice.dto.MatchDetailDto;
-import org.thingai.app.scoringservice.entity.score.Score;
+import org.thingai.app.scoringservice.entity.Score;
+import org.thingai.app.scoringservice.entity.RankingEntry;
+import org.thingai.app.scoringservice.entity.RankingStat;
 
 public interface IRankingStrategy {
     RankingEntry[] sortRankingEntries(RankingEntry[] entries);
